layers service renamed. App and env params added
